3주차_로또_re/src/test/java/lotto/domainTest/LottoTest.java
package lotto.domainTest;

import lotto.domain.Lotto;
import lotto.enums.GameResult;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;

class LottoTest {
    @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
    @Test
    void createLottoByOverSize() {
        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
                .isInstanceOf(IllegalArgumentException.class);
    }

    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
    @Test
    void createLottoByDuplicatedNumber() {
        // TODO: 이 테스트가 통과할 수 있게 구현 코드 작성
        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
                .isInstanceOf(IllegalArgumentException.class);
    }

    @DisplayName("보너스 넘버가 이미 있는 경우 참을 반환한다. ")
    @Test
    void containsBonusNumberTest(){
        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
        assertThat(lotto.containsBonusNumber(6)).isTrue();
    }

    @DisplayName("같은 개수에 따라 GameResult형으로 반환한다. ")
    @Test
    void matchWithPlayerLottoTest(){
        Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
        assertThat(lotto.matchWithPlayerLotto(List.of(1, 2, 3, 4, 5, 6)))
                .isEqualTo(GameResult.SIX);
    }
}
